refactor(team): extract OpenCode runtime member session acceptance from the check-in flow
TeamProvisioningOpenCodeRuntimeCheckin.ts sits at 799 lines against the
800-line source size guard. The member session acceptance rule inside it
is self-contained: it answers a single question - may this runtime
check-in claim this member identity - and it has two entry points, one
that reads the persisted state itself and one that decides from state a
caller already holds.

Move that rule into TeamProvisioningOpenCodeRuntimeMemberSessionAcceptance.ts.
The moved bodies are byte-identical; the only edits are the `export`
keyword on the previously file-private state-based entry point and the
re-export that keeps the check-in module's public surface unchanged, so
TeamProvisioningOpenCodeRuntimeDelivery.ts keeps importing
assertOpenCodeRuntimeMemberSessionAccepted from where it always did.

The extraction comes first, on its own, because the follow-up fix in this
series adds lines back to the check-in module. Landing it as a separate
commit means no intermediate tree in this branch ever crosses the size
guard, and the behavior change that follows is reviewable as a diff
against a file that is already the right size rather than as a diff
tangled up with a 115-line move.
